requirements.txt
backports.csv==1.0.7
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.1
django-allauth==0.40.0
django-ckeditor==5.7.1
django-contrib-comments==1.9.1
django-cors-headers==3.0.2
django-crispy-forms==1.7.2
django-filter==2.2.0
django-formtools==2.1
django-haystack==2.8.1
django-import-export==1.2.0
django-js-asset==1.2.2
django-mdeditor==0.1.16
django-model-utils==3.2.0
django-mptt==0.10.0
django-notifications-hq==1.5.0
django-redis==4.10.0
django-rest-swagger==2.2.0
django-reversion==3.0.4
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
drf-extensions==0.5.0
et-xmlfile==1.0.1
future==0.17.1
httplib2==0.9.2
idna==2.8
itypes==1.1.0
jdcal==1.4.1
jieba==0.39
Jinja2==2.10.1
jsonfield==2.0.2
Markdown==3.1.1
MarkupSafe==1.1.1
mysqlclient==1.4.2.post1
oauthlib==3.1.0
odfpy==1.4.0
openapi-codec==1.3.2
openpyxl==2.6.2
Pillow==6.1.0
Pygments==2.4.2
PyJWT==1.7.1
PyMySQL==0.9.3
python3-openid==3.1.0
pytz==2019.1
PyYAML==5.1.2
redis==3.3.8
requests==2.22.0
requests-oauthlib==1.2.0
simplejson==3.16.0
six==1.12.0
social-auth-core==3.2.0
sqlparse==0.3.0
tablib==0.13.0
uritemplate==3.0.0
urllib3==1.25.3
Whoosh==2.7.4
wincertstore==0.2
xadmin==2.0.1
xlrd==1.2.0
xlwt==1.3.0
